Moved existing solutions into a more specific folder
